requirements.txt
PyGithub~=1.51